// apps/web/src/generated/SessionSummary.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { SessionKind } from "./SessionKind";
import type { SessionStatus } from "./SessionStatus";

export type SessionSummary = { id: number, kind: SessionKind, unit_id: string | null, status: SessionStatus, started_at: string, ended_at: string | null, };